requirements.txt
aiomysql==0.2.0
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.2
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
exceptiongroup==1.2.0
fastapi==0.105.0
greenlet==3.0.2
h11==0.14.0
idna==3.6
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
PyMySQL==1.1.0
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3.post1
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tortoise-orm==0.20.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
Werkzeug==3.0.1
